order-system/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
gspread==5.12.0
oauth2client==4.1.3
mysql-connector-python==8.2.0
pandas==2.1.3
pytz==2023.3
python-dotenv==1.0.0 